move camera logic to presenter
