+++ app/src/main/java/org/wit/archaeologicalfieldwork/views/hillfort/HillfortFragment.kt
@@ -3,8 +3,6 @@ package org.wit.archaeologicalfieldwork.views.hillfort
 import android.content.Intent
 import android.graphics.Bitmap
 import android.os.Bundle
-import android.os.Environment
-import android.provider.MediaStore
 import com.google.android.material.textfield.TextInputEditText
 import androidx.fragment.app.Fragment
 import androidx.appcompat.app.AppCompatActivity
@@ -19,7 +17,6 @@ import com.google.android.gms.maps.MapView
 import com.google.android.gms.maps.model.Marker
 
 import org.jetbrains.anko.AnkoLogger
-import org.jetbrains.anko.info
 import org.jetbrains.anko.support.v4.intentFor
 import org.jetbrains.anko.support.v4.toast
 import org.wit.archaeologicalfieldwork.R
@@ -27,10 +24,6 @@ import org.wit.archaeologicalfieldwork.models.data.DataModel
 import org.wit.archaeologicalfieldwork.models.hillfort.HillfortModel
 import org.wit.archaeologicalfieldwork.models.user.UserModel
 import org.wit.archaeologicalfieldwork.views.hillfortlist.HillfortListActivity
-import java.io.File
-import java.io.IOException
-import java.text.SimpleDateFormat
-import java.util.Date
 
 class HillfortFragment : Fragment(), AnkoLogger, GoogleMap.OnMarkerDragListener {
 
@@ -46,7 +39,6 @@ class HillfortFragment : Fragment(), AnkoLogger, GoogleMap.OnMarkerDragListener 
     lateinit var ratingBar: RatingBar
     lateinit var cameraBtn: Button
     lateinit var imageView: ImageView
-    lateinit var currentPhotoPath: String
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         presenter = HillfortPresenter(this)
@@ -102,7 +94,7 @@ class HillfortFragment : Fragment(), AnkoLogger, GoogleMap.OnMarkerDragListener 
         }
 
         cameraBtn?.setOnClickListener {
-            doCamera()
+            presenter.doCamera()
         }
 
         return view
@@ -110,36 +102,6 @@ class HillfortFragment : Fragment(), AnkoLogger, GoogleMap.OnMarkerDragListener 
 
     fun handleButton(button: Button?, view: View) {
         if (!presenter.edit()) button?.visibility = View.INVISIBLE
-    }
-
-    fun doCamera() {
-        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also { takePictureIntent ->
-            takePictureIntent.resolveActivity(activity!!.packageManager)?.also {
-                val photoFile: File? = try {
-                    createImageFile()
-                } catch (ex: IOException) {
-                    null
-                }
-                photoFile?.also {
-                    startActivityForResult(takePictureIntent, IMAGE_CAPTURE)
-                }
-            }
-        }
-    }
-
-    @Throws(IOException::class)
-    private fun createImageFile(): File {
-        // Create an image file name
-        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        val storageDir: File = activity!!.applicationContext.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-        return File.createTempFile(
-            "JPEG_${timeStamp}_", /* prefix */
-            ".jpg", /* suffix */
-            storageDir /* directory */
-        ).apply {
-            // Save a file: path for use with ACTION_VIEW intents
-            currentPhotoPath = absolutePath
-        }
     }
 
     companion object {
@@ -173,7 +135,7 @@ class HillfortFragment : Fragment(), AnkoLogger, GoogleMap.OnMarkerDragListener 
             }
             IMAGE_CAPTURE -> {
                 val imageBitmap = data!!.extras.get("data") as Bitmap
-                presenter.doUploadBitmap(imageBitmap, currentPhotoPath, hillfort)
+                presenter.doUploadBitmap(imageBitmap, hillfort)
             }
         }
     }
