+++ app/src/main/java/org/wit/archaeologicalfieldwork/views/hillfort/HillfortPresenter.kt
@@ -1,7 +1,10 @@
 package org.wit.archaeologicalfieldwork.views.hillfort
 
 import android.annotation.SuppressLint
+import android.content.Intent
 import android.graphics.Bitmap
+import android.os.Environment
+import android.provider.MediaStore
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
@@ -12,7 +15,6 @@ import com.google.android.gms.maps.model.MarkerOptions
 import kotlinx.coroutines.experimental.android.UI
 import kotlinx.coroutines.experimental.async
 import org.jetbrains.anko.AnkoLogger
-import org.jetbrains.anko.info
 import org.wit.archaeologicalfieldwork.helpers.checkLocationPermissions
 import org.wit.archaeologicalfieldwork.helpers.getDate
 import org.wit.archaeologicalfieldwork.helpers.isPermissionGranted
@@ -22,6 +24,10 @@ import org.wit.archaeologicalfieldwork.models.data.DataModel
 import org.wit.archaeologicalfieldwork.models.location.Location
 import org.wit.archaeologicalfieldwork.models.stats.StatsModel
 import org.wit.archaeologicalfieldwork.models.user.UserModel
+import java.io.File
+import java.io.IOException
+import java.text.SimpleDateFormat
+import java.util.Date
 
 class HillfortPresenter(val view: HillfortFragment) : AnkoLogger {
 
@@ -29,6 +35,8 @@ class HillfortPresenter(val view: HillfortFragment) : AnkoLogger {
 
     var defaultLocation = Location(52.245696, -7.139102, 15f)
     var locationService: FusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(view.activity!!)
+
+    lateinit var currentPhotoPath: String
 
     var hillfort = DataModel()
     var stat = StatsModel()
@@ -119,10 +127,37 @@ class HillfortPresenter(val view: HillfortFragment) : AnkoLogger {
     }
 
     fun doCamera() {
+        Intent(MediaStore.ACTION_IMAGE_CAPTURE).also { takePictureIntent ->
+            takePictureIntent.resolveActivity(view.activity!!.packageManager)?.also {
+                val photoFile: File? = try {
+                    createImageFile()
+                } catch (ex: IOException) {
+                    null
+                }
+                photoFile?.also {
+                    view.startActivityForResult(takePictureIntent, view.IMAGE_CAPTURE)
+                }
+            }
+        }
     }
 
-    fun doUploadBitmap(bitmap: Bitmap, name: String, hillfort: DataModel) {
-        fireStore.updateBitMapImage(bitmap, name, hillfort)
+    fun doUploadBitmap(bitmap: Bitmap, hillfort: DataModel) {
+        fireStore.updateBitMapImage(bitmap, currentPhotoPath, hillfort)
+    }
+
+    @Throws(IOException::class)
+    private fun createImageFile(): File {
+        // Create an image file name
+        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
+        val storageDir: File = view.activity!!.applicationContext.getExternalFilesDir(Environment.DIRECTORY_PICTURES)
+        return File.createTempFile(
+            "JPEG_${timeStamp}_", /* prefix */
+            ".jpg", /* suffix */
+            storageDir /* directory */
+        ).apply {
+            // Save a file: path for use with ACTION_VIEW intents
+            currentPhotoPath = absolutePath
+        }
     }
 
     fun doVisit(user: UserModel) {
